server/requirements.txt
Flask
Flask-MySQL
mysql-connector-python
flask-cors